+++ app/src/main/java/com/github/damontecres/stashapp/MainFragment.kt
@@ -35,7 +35,7 @@ import com.github.damontecres.stashapp.api.ConfigurationQuery
 import com.github.damontecres.stashapp.api.type.FilterMode
 import com.github.damontecres.stashapp.api.type.FindFilterType
 import com.github.damontecres.stashapp.api.type.SortDirectionEnum
-import com.github.damontecres.stashapp.presenters.stashPresenterSelector
+import com.github.damontecres.stashapp.presenters.StashPresenter
 import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.launch
 import java.util.Timer
@@ -115,9 +115,10 @@ class MainFragment : BrowseSupportFragment() {
     }
 
     private fun prepareBackgroundManager() {
-
         mBackgroundManager = BackgroundManager.getInstance(activity)
         mBackgroundManager.attach(requireActivity().window)
+        mBackgroundManager.color =
+            ContextCompat.getColor(requireContext(), R.color.default_background)
         mDefaultBackground =
             ContextCompat.getDrawable(requireActivity(), R.drawable.default_background)
         mMetrics = DisplayMetrics()
@@ -232,7 +233,7 @@ class MainFragment : BrowseSupportFragment() {
                             val frontPageContent =
                                 (ui as Map<String, *>)["frontPageContent"] as List<Map<String, *>>
                             for (frontPageFilter: Map<String, *> in frontPageContent) {
-                                val adapter = ArrayObjectAdapter(stashPresenterSelector)
+                                val adapter = ArrayObjectAdapter(StashPresenter.SELECTOR)
                                 adapters.add(adapter)
 
                                 val filterType = frontPageFilter["__typename"] as String
