+++ app/src/main/java/com/github/damontecres/stashapp/presenters/PerformerPresenter.kt
@@ -13,49 +13,17 @@ import com.github.damontecres.stashapp.api.fragment.PerformerData
 import com.github.damontecres.stashapp.createGlideUrl
 import kotlin.properties.Delegates
 
-/**
- * A CardPresenter is used to generate Views and bind Objects to them on demand.
- * It contains an ImageCardView.
- */
-class PerformerPresenter : Presenter() {
-    private var vParent: ViewGroup by Delegates.notNull()
-    private var mDefaultCardImage: Drawable? = null
-    private var sSelectedBackgroundColor: Int by Delegates.notNull()
-    private var sDefaultBackgroundColor: Int by Delegates.notNull()
-
-    override fun onCreateViewHolder(parent: ViewGroup): ViewHolder {
-        Log.d(TAG, "onCreateViewHolder")
-        vParent = parent
-        sDefaultBackgroundColor = ContextCompat.getColor(parent.context, R.color.default_background)
-        sSelectedBackgroundColor =
-            ContextCompat.getColor(parent.context, R.color.selected_background)
-        mDefaultCardImage =
-            ContextCompat.getDrawable(parent.context, R.drawable.baseline_camera_indoor_48)
-
-        val cardView = object : ImageCardView(parent.context) {
-            override fun setSelected(selected: Boolean) {
-                updateCardBackgroundColor(this, selected)
-                super.setSelected(selected)
-            }
-        }
-
-        cardView.isFocusable = true
-        cardView.isFocusableInTouchMode = true
-        updateCardBackgroundColor(cardView, false)
-        return ViewHolder(cardView)
-    }
+class PerformerPresenter : StashPresenter() {
 
     override fun onBindViewHolder(viewHolder: ViewHolder, item: Any?) {
         val performer = item as PerformerData
         val cardView = viewHolder.view as ImageCardView
 
-        Log.d(TAG, "onBindViewHolder")
         if (performer.image_path != null) {
             val title =
                 performer.name + (if (!performer.disambiguation.isNullOrBlank()) " (${performer.disambiguation})" else "")
             cardView.titleText = title
             cardView.contentText = "${performer.scene_count} Scenes"
-//            cardView.contentText = movie.studio
             cardView.setMainImageDimensions(CARD_WIDTH, CARD_HEIGHT)
             val apiKey = PreferenceManager.getDefaultSharedPreferences(vParent.context)
                 .getString("stashApiKey", "")
@@ -68,27 +36,11 @@ class PerformerPresenter : Presenter() {
         }
     }
 
-    override fun onUnbindViewHolder(viewHolder: ViewHolder) {
-        Log.d(TAG, "onUnbindViewHolder")
-        val cardView = viewHolder.view as ImageCardView
-        // Remove references to images so that the garbage collector can free up memory
-        cardView.badgeImage = null
-        cardView.mainImage = null
-    }
-
-    private fun updateCardBackgroundColor(view: ImageCardView, selected: Boolean) {
-        val color = if (selected) sSelectedBackgroundColor else sDefaultBackgroundColor
-        // Both background colors should be set because the view"s background is temporarily visible
-        // during animations.
-        view.setBackgroundColor(color)
-        view.setInfoAreaBackgroundColor(color)
-    }
-
     companion object {
-        private val TAG = "CardPresenter"
+        private const val TAG = "CardPresenter"
 
-        private val CARD_HEIGHT = 381
-        private val CARD_WIDTH = CARD_HEIGHT * 2 / 3
+        private const val CARD_HEIGHT = 381
+        private const val CARD_WIDTH = CARD_HEIGHT * 2 / 3
 
     }
 }
