+++ app/src/main/java/com/github/damontecres/stashapp/presenters/ScenePresenter.kt
@@ -14,46 +14,11 @@ import com.github.damontecres.stashapp.createGlideUrl
 import java.io.File
 import kotlin.properties.Delegates
 
-
-/**
- * A CardPresenter is used to generate Views and bind Objects to them on demand.
- * It contains an ImageCardView.
- */
-class ScenePresenter : Presenter() {
-    private var vParent: ViewGroup by Delegates.notNull()
-    private var mDefaultCardImage: Drawable? = null
-    private var sSelectedBackgroundColor: Int by Delegates.notNull()
-    private var sDefaultBackgroundColor: Int by Delegates.notNull()
-
-    override fun onCreateViewHolder(parent: ViewGroup): Presenter.ViewHolder {
-        vParent = parent
-
-        sDefaultBackgroundColor = ContextCompat.getColor(parent.context, R.color.default_background)
-        sSelectedBackgroundColor = ContextCompat.getColor(
-            parent.context,
-            R.color.selected_background
-        )
-        mDefaultCardImage =
-            ContextCompat.getDrawable(parent.context, R.drawable.baseline_camera_indoor_48)
-
-        val cardView = object : ImageCardView(parent.context) {
-            override fun setSelected(selected: Boolean) {
-                updateCardBackgroundColor(this, selected)
-                super.setSelected(selected)
-            }
-        }
-
-        cardView.isFocusable = true
-        cardView.isFocusableInTouchMode = true
-        updateCardBackgroundColor(cardView, false)
-        return ViewHolder(cardView)
-    }
+class ScenePresenter : StashPresenter() {
 
     override fun onBindViewHolder(viewHolder: ViewHolder, item: Any?) {
-//        val scene = sceneFromSlimSceneData(item as SlimSceneData)
         val scene = item as SlimSceneData
         val cardView = viewHolder.view as ImageCardView
-        Log.d(TAG, "onBindViewHolder: ${scene.title}")
         if (scene.title.isNullOrBlank()) {
             val path = scene.files.firstOrNull()?.videoFileData?.path
             if (path != null) {
@@ -78,24 +43,8 @@ class ScenePresenter : Presenter() {
         }
     }
 
-    override fun onUnbindViewHolder(viewHolder: Presenter.ViewHolder) {
-        Log.d(TAG, "onUnbindViewHolder")
-        val cardView = viewHolder.view as ImageCardView
-        // Remove references to images so that the garbage collector can free up memory
-        cardView.badgeImage = null
-        cardView.mainImage = null
-    }
-
-    private fun updateCardBackgroundColor(view: ImageCardView, selected: Boolean) {
-        val color = if (selected) sSelectedBackgroundColor else sDefaultBackgroundColor
-        // Both background colors should be set because the view"s background is temporarily visible
-        // during animations.
-        view.setBackgroundColor(color)
-        view.setInfoAreaBackgroundColor(color)
-    }
-
     companion object {
-        private val TAG = "CardPresenter"
+        private val TAG = "ScenePresenter"
 
         const val CARD_WIDTH = 351
         const val CARD_HEIGHT = 198
