+++ app/src/main/java/com/github/damontecres/stashapp/presenters/StashPresenter.kt
@@ -1,0 +1,75 @@
+package com.github.damontecres.stashapp.presenters
+
+import android.graphics.drawable.Drawable
+import android.util.Log
+import android.view.ViewGroup
+import androidx.core.content.ContextCompat
+import androidx.leanback.widget.ClassPresenterSelector
+import androidx.leanback.widget.ImageCardView
+import androidx.leanback.widget.Presenter
+import androidx.preference.PreferenceManager
+import com.bumptech.glide.Glide
+import com.github.damontecres.stashapp.R
+import com.github.damontecres.stashapp.api.fragment.PerformerData
+import com.github.damontecres.stashapp.api.fragment.SlimSceneData
+import com.github.damontecres.stashapp.api.fragment.StudioData
+import com.github.damontecres.stashapp.createGlideUrl
+import com.github.damontecres.stashapp.data.Tag
+import java.io.File
+import kotlin.properties.Delegates
+
+
+abstract class StashPresenter : Presenter() {
+    protected var vParent: ViewGroup by Delegates.notNull()
+    protected var mDefaultCardImage: Drawable? = null
+    private var sSelectedBackgroundColor: Int by Delegates.notNull()
+    private var sDefaultBackgroundColor: Int by Delegates.notNull()
+
+    override fun onCreateViewHolder(parent: ViewGroup): Presenter.ViewHolder {
+        vParent = parent
+
+        sDefaultBackgroundColor =
+            ContextCompat.getColor(parent.context, R.color.default_card_background)
+        sSelectedBackgroundColor = ContextCompat.getColor(
+            parent.context,
+            R.color.selected_background
+        )
+        mDefaultCardImage =
+            ContextCompat.getDrawable(parent.context, R.drawable.baseline_camera_indoor_48)
+
+        val cardView = object : ImageCardView(parent.context) {
+            override fun setSelected(selected: Boolean) {
+                updateCardBackgroundColor(this, selected)
+                super.setSelected(selected)
+            }
+        }
+
+        cardView.isFocusable = true
+        cardView.isFocusableInTouchMode = true
+        updateCardBackgroundColor(cardView, false)
+        return ViewHolder(cardView)
+    }
+
+    override fun onUnbindViewHolder(viewHolder: Presenter.ViewHolder) {
+        val cardView = viewHolder.view as ImageCardView
+        // Remove references to images so that the garbage collector can free up memory
+        cardView.badgeImage = null
+        cardView.mainImage = null
+    }
+
+    private fun updateCardBackgroundColor(view: ImageCardView, selected: Boolean) {
+        val color = if (selected) sSelectedBackgroundColor else sDefaultBackgroundColor
+        // Both background colors should be set because the view"s background is temporarily visible
+        // during animations.
+        view.setBackgroundColor(color)
+        view.setInfoAreaBackgroundColor(color)
+    }
+
+    companion object {
+        val SELECTOR: ClassPresenterSelector = ClassPresenterSelector()
+            .addClassPresenter(PerformerData::class.java, PerformerPresenter())
+            .addClassPresenter(SlimSceneData::class.java, ScenePresenter())
+            .addClassPresenter(StudioData::class.java, StudioPresenter())
+            .addClassPresenter(Tag::class.java, TagPresenter())
+    }
+}
